requirements.txt
alembic==1.2.1
aniso8601==8.0.0
attrs==19.2.0
bcrypt==3.1.7
bson==0.5.8
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.2
six==1.12.0
Werkzeug==0.16.0
